Reorganize gruntfile.
